chore: Refactor MOTD loading functionality and clean up legacy code
- Completed the refactor of the `load_motd()` function, moving it from the deprecated `server/real_time.py` to a new `server/utils/motd_loader.py` module, enhancing modularity and maintainability.
- Updated the associated test suite, creating `test_motd_loader.py` to ensure comprehensive coverage of the new MOTD loading functionality, including file handling and error management.
- Removed the legacy `server/real_time.py` file, which contained only the outdated `load_motd()` function, streamlining the codebase and eliminating unnecessary components.

This commit contributes to a cleaner and more organized code structure in MythosMUD, aligning with our ongoing efforts to refine development practices and improve code quality.
